fix: 검색시 redis에 저장된 stock데이터로 값불러오기 (#67)

+++ src/main/java/back/whats_your_ETF/dto/StockRankResponse.java
@@ -3,6 +3,6 @@ package back.whats_your_ETF.dto;
 public record StockRankResponse (
         String stockCode,
         String stockName,
-        Double price,
+        Long price,
         String priceChangeRate
 ) {}
